seperated header and cpp files, no issues with compilation, memory segmentation error?
