+++ .gigitnore
@@ -1,1 +1,3 @@
 .vscode/
+Project-1/*.gch
+Project-1/*.o
