+++ Project-1/ProcessControlBlock.cpp
@@ -1,27 +1,24 @@
 #include <iostream>
 #include <fstream>
 #include <vector>
-
-#include"Process.cpp"
+#include "Process.h"
+#include "ProcessControlBlock.h"
 using namespace std;
 
-class ProcessControlBlock {
-    public:
-        vector<Process> processes;
+ProcessControlBlock::ProcessControlBlock() {};
 
-        void readFromFile(string path) {
-            string lineText;
-            ifstream datafileIfstream(path);
-            while (getline(datafileIfstream, lineText)) {
-                Process newProcess;
-                newProcess.loadFromLine(lineText);
-                processes.insert(processes.end() + 1, newProcess);
-            }
-        }
+void ProcessControlBlock::readFromFile(string path) {
+    string lineText;
+    ifstream datafileIfstream(path);
+    while (getline(datafileIfstream, lineText)) {
+        Process newProcess;
+        newProcess.loadFromLine(lineText);
+        processes.insert(processes.end() + 1, newProcess);
+    }
+}
 
-        void printProcesses() {
-            for (int i = 0; i < processes.size(); i++) {
-                cout << processes[i].testString << "\n";
-            };
-        }
-};
+void ProcessControlBlock::printProcesses() {
+    for (int i = 0; i < processes.size(); i++) {
+        cout << processes[i].testString << "\n";
+    };
+}
