+++ Project-1/main.cpp
@@ -1,17 +1,19 @@
 #include <iostream>
 #include <fstream>
 #include <vector>
-
-#include"Process.cpp"
-#include"ProcessControlBlock.cpp"
+#include "Process.h"
+#include "ProcessControlBlock.h"
 using namespace std;
 
-main() {
-    cout << "Please enter the file path to the file you would like to use";
+int main() {
+    cout << "Please enter the file path to the file you would like to use\n\n";
     string path;
     cin >> path;
-    
+    cout << path;
+
     ProcessControlBlock PCB;
-    PCB.readFromFile(path);
+    PCB.readFromFile("../test.txt");
     PCB.printProcesses();
+
+    return 0;
 };
